steganography_hunter_agent.py
# Copyright 2025 Amazon.com and its affiliates; all rights reserved.
# This file is Amazon Web Services Content and may not be distributed without permission.

"""
Steganography Hunter Agent - Hidden Message Detection
Handles acrostic messages, spacing patterns, and numeric encoding schemes using AI agents
"""

from strands import Agent, tool
from strands.models import BedrockModel

# Initialize the model
model_id = "us.amazon.nova-pro-v1:0"
model = BedrockModel(model_id=model_id)

@tool
def acrostic_detector(text: str) -> str:
    """Detect acrostic messages using an AI agent"""
    try:
        ACROSTIC_AGENT_PROMPT = """
        You are a cryptography expert specializing in acrostic message detection.

        {Todo}        

        Always return your answer in the format: "DECODED: [result]" or "NO HIDDEN MESSAGE DETECTED"
        """
        
        acrostic_agent = Agent(
            system_prompt=ACROSTIC_AGENT_PROMPT,
            model=model
        )
        response = acrostic_agent(f"Detect acrostic messages in this text: {text}")
        return str(response)
    except Exception as e:
        return f"Error detecting acrostic messages: {str(e)}"

@tool
def numeric_encoding_decoder(text: str) -> str:
    """Decode numeric encoding schemes using an AI agent"""
    try:
        NUMERIC_AGENT_PROMPT = """
        You are a cryptography expert specializing in numeric encoding schemes.
        
        Common numeric encoding methods:
        {TODO I know this one, i learnt it in boyscouts.  You gotta give it a list examples of A1Z26, ASCII, phone keypad, and some other number things}


        Try all applicable methods and return the most readable result.
        
        Always return your answer in the format: "DECODED: [result]" or "NO VALID ENCODING DETECTED"
        """
        
        numeric_agent = Agent(
            system_prompt=NUMERIC_AGENT_PROMPT,
            model=model
        )
        response = numeric_agent(f"Decode numeric encoding in this text: {text}")
        return str(response)
    except Exception as e:
        return f"Error decoding numeric encoding: {str(e)}"

